feat(gabinet): surface dated schedule periods on the timetable page (#331)
Extract FlexibleScheduleEditor and groupSchedulesIntoPeriods into a shared
component so the timetable page can manage dated multi-period overrides — not
just the default-period schedule. The timetable row now resolves the period
active for today (dated overrides take priority over the default), shows an
"Active period" range badge and an "N dated periods" badge to indicate the
employee has additional overrides, and opens the full editor in the side sheet.
